# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type QuestionnaireInputProductType =
  (typeof QuestionnaireInputProductType)[keyof typeof QuestionnaireInputProductType];

export const QuestionnaireInputProductType = {
  deodorant: "deodorant",
  shower_gel: "shower_gel",
} as const;

export type QuestionnaireInputSkinType =
  (typeof QuestionnaireInputSkinType)[keyof typeof QuestionnaireInputSkinType];

export const QuestionnaireInputSkinType = {
  dry: "dry",
  oily: "oily",
  combination: "combination",
  normal: "normal",
  sensitive: "sensitive",
} as const;

export type QuestionnaireInputSweatLevel =
  | (typeof QuestionnaireInputSweatLevel)[keyof typeof QuestionnaireInputSweatLevel]
  | null;

export const QuestionnaireInputSweatLevel = {
  light: "light",
  moderate: "moderate",
  heavy: "heavy",
} as const;

export type QuestionnaireInputSkinIssuesItem =
  (typeof QuestionnaireInputSkinIssuesItem)[keyof typeof QuestionnaireInputSkinIssuesItem];

export const QuestionnaireInputSkinIssuesItem = {
  redness: "redness",
  itching: "itching",
  dryness: "dryness",
  irritation: "irritation",
  darkening: "darkening",
  acne: "acne",
} as const;

export type QuestionnaireInputAllergiesItem =
  (typeof QuestionnaireInputAllergiesItem)[keyof typeof QuestionnaireInputAllergiesItem];

export const QuestionnaireInputAllergiesItem = {
  fragrances: "fragrances",
  alcohol: "alcohol",
  parabens: "parabens",
  sulfates: "sulfates",
  lanolin: "lanolin",
  propylene_glycol: "propylene_glycol",
} as const;

export type QuestionnaireInputLifestyle =
  (typeof QuestionnaireInputLifestyle)[keyof typeof QuestionnaireInputLifestyle];

export const QuestionnaireInputLifestyle = {
  active: "active",
  office: "office",
  outdoor: "outdoor",
  mixed: "mixed",
} as const;

export type QuestionnaireInputFragranceStyle =
  (typeof QuestionnaireInputFragranceStyle)[keyof typeof QuestionnaireInputFragranceStyle];

export const QuestionnaireInputFragranceStyle = {
  unscented: "unscented",
  fresh: "fresh",
  floral: "floral",
  woody: "woody",
  citrus: "citrus",
  herbal: "herbal",
} as const;

export type QuestionnaireInputTexturePreference =
  (typeof QuestionnaireInputTexturePreference)[keyof typeof QuestionnaireInputTexturePreference];

export const QuestionnaireInputTexturePreference = {
  gel: "gel",
  cream: "cream",
  spray: "spray",
  stick: "stick",
  foam: "foam",
} as const;

export interface QuestionnaireInput {
  productType: QuestionnaireInputProductType;
  skinType: QuestionnaireInputSkinType;
  sweatLevel?: QuestionnaireInputSweatLevel;
  skinIssues?: QuestionnaireInputSkinIssuesItem[];
  allergies?: QuestionnaireInputAllergiesItem[];
  lifestyle: QuestionnaireInputLifestyle;
  fragranceStyle: QuestionnaireInputFragranceStyle;
  texturePreference: QuestionnaireInputTexturePreference;
}

export interface FormulaIngredient {
  name: string;
  purpose: string;
  percentage?: string | null;
}

export interface FormulaSection {
  title: string;
  description?: string | null;
  ingredients: FormulaIngredient[];
}

export type FormulationResultProductType =
  (typeof FormulationResultProductType)[keyof typeof FormulationResultProductType];

export const FormulationResultProductType = {
  deodorant: "deodorant",
  shower_gel: "shower_gel",
} as const;

export interface FormulationResult {
  id: string;
  createdAt: string;
  productType: FormulationResultProductType;
  productName: string;
  tagline: string;
  baseFormula: FormulaSection;
  odorControl: FormulaSection;
  skinCareActives: FormulaSection;
  fragranceProfile: FormulaSection;
  applicationInstructions: string;
  irishWaterNote?: string | null;
}