# lib/api-zod/src/generated/types/formulaIngredient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FormulaIngredient {
  name: string;
  purpose: string;
  percentage?: string | null;
}