# lib/api-zod/src/generated/types/formulaSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FormulaIngredient } from "./formulaIngredient";

export interface FormulaSection {
  title: string;
  description?: string | null;
  ingredients: FormulaIngredient[];
}