# lib/api-zod/src/generated/types/formulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FormulaSection } from "./formulaSection";
import type { FormulationResultProductType } from "./formulationResultProductType";

export interface FormulationResult {
  id: string;
  createdAt: string;
  productType: FormulationResultProductType;
  productName: string;
  tagline: string;
  baseFormula: FormulaSection;
  odorControl: FormulaSection;
  skinCareActives: FormulaSection;
  fragranceProfile: FormulaSection;
  applicationInstructions: string;
  irishWaterNote?: string | null;
}