# lib/api-zod/src/generated/types/questionnaireInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionnaireInputAllergiesItem } from "./questionnaireInputAllergiesItem";
import type { QuestionnaireInputFragranceStyle } from "./questionnaireInputFragranceStyle";
import type { QuestionnaireInputLifestyle } from "./questionnaireInputLifestyle";
import type { QuestionnaireInputProductType } from "./questionnaireInputProductType";
import type { QuestionnaireInputSkinIssuesItem } from "./questionnaireInputSkinIssuesItem";
import type { QuestionnaireInputSkinType } from "./questionnaireInputSkinType";
import type { QuestionnaireInputSweatLevel } from "./questionnaireInputSweatLevel";
import type { QuestionnaireInputTexturePreference } from "./questionnaireInputTexturePreference";

export interface QuestionnaireInput {
  productType: QuestionnaireInputProductType;
  skinType: QuestionnaireInputSkinType;
  sweatLevel?: QuestionnaireInputSweatLevel;
  skinIssues?: QuestionnaireInputSkinIssuesItem[];
  allergies?: QuestionnaireInputAllergiesItem[];
  lifestyle: QuestionnaireInputLifestyle;
  fragranceStyle: QuestionnaireInputFragranceStyle;
  texturePreference: QuestionnaireInputTexturePreference;
}